requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Document Generation
python-docx>=0.8.11

# ML Enhancement
imbalanced-learn>=0.11.0

# Data Handling
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
